+++ src/main/java/joshie/harvest/tools/item/ItemSickle.java
@@ -100,7 +100,6 @@ public class ItemSickle extends ItemTool<ItemHoe> implements IBreakCrops {
                 return 0F;
             }
 
-
             ToolTier tier = getTier(stack);
             //Facing North, We Want East and West to be 1, left * this.left
             for (int x2 = getXMinus(tier, front, pos.getX()); x2 <= getXPlus(tier, front, pos.getX()); x2++) {