+++ src/main/resources/changelog.md
@@ -18,6 +18,8 @@
 * Moved the sprinkler to the carpenters store
 * Fodder now costs 100 gold, but will last for 10 feeds
 * Chicken Feed now costs 50 gold, but will last for 10 feeds
+#Tools
+* Tools will have [BROKEN] in front of their name, when they can't be used
 #Other
 * Crops will now be plantable on more than farmland,
     however they will only grow on farmland when a mod adds support